Refactor CLI Folder Structure (#12)
- Creates separate command folder structure

- Refactored cli __init__ to import commands explicitly
