+++ RELEASE.md
@@ -1,0 +1,3 @@
+Release type: minor
+
+Refactored CLI folder structure, importing click commands from a subfolder. Follows click's complex example.
